refactor: now database is a model
