--- www/api/v1.3/bin/refresh.sh
@@ -1,11 +1,0 @@
-#!/usr/bin/env bash
-
-set -o errexit
-
-# shellcheck source=../../../../bin/cr-lib
-. "$(dirname "${BASH_SOURCE[0]}")"/../../../../bin/cr-lib
-
-SWD="$(dirname "${BASH_SOURCE[0]}")"
-V="$(basename "$(dirname "$SWD")")"
-
-curl -fsSL "http://${CR_HTTP_HOST}/api/$V/admin/database?force=1&quiet=true"
